lncli: broke up cli command unit tests into a file per command
